optimized_bot.py
import os
from web3 import Web3
from dotenv import load_dotenv
load_dotenv(os.path.expanduser('~/.secrets/flashloan.env'))

# --- Elite-Grade Optimizations ---
# 1. Async execution (aiohttp)
# 2. Error handling with retries
# 3. Gas price optimization
# 4. Arbitrage opportunity scoring
