requirements.txt
dash==2.7.1
dash-bootstrap-components==1.3.1
flask>=2.3.3
streamlit>=1.32.0
pandas==1.5.3
numpy==1.24.3
plotly==5.11.0
waitress==2.1.2
protobuf==3.20.3

psycopg2-binary>=2.9.6
flask>=3.1.1
gunicorn
857f1ab (Add untracked operational files and configuration assets)
